added mapping between layers

+++ data/src/main/java/com/amaringo/data/carecenter/network/model/Location.kt
@@ -1,6 +1,6 @@
 package com.amaringo.data.carecenter.network.model
 
 data class Location(
-    val latitude: Int,
-    val longitude: Int
+    val latitude: Double,
+    val longitude: Double
 )
